requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
chex==0.1.90
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
dm-haiku==0.0.14
etils==1.13.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flax==0.11.1
fonttools==4.59.0
fqdn==1.5.1
fsspec==2025.7.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
identify==2.6.13
idna==3.10
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jax==0.7.0
jax-cuda12-pjrt==0.7.0
jax-cuda12-plugin==0.7.0
jaxlib==0.7.0
jedi==0.19.2
Jinja2==3.1.6
jmp==0.0.4
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
lark==1.2.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
meson==1.8.3
mistune==3.1.3
ml_dtypes==0.5.3
msgpack==1.1.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.2
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.12.0.46
nvidia-cufft-cu12==11.4.1.4
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.27.7
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.3.20
opt_einsum==3.4.0
optax==0.2.5
orbax-checkpoint==0.11.21
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pre_commit==4.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
pyscf==2.10.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.0
scipy==1.16.1
Send2Trash==1.8.3
setuptools==78.1.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
tabulate==0.9.0
tensorstore==0.1.76
terminado==0.18.1
tinycss2==1.4.0
toolz==1.0.0
tornado==6.5.2
traitlets==5.14.3
treescope==0.1.10
types-python-dateutil==2.9.0.20250809
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
virtualenv==20.33.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
zipp==3.23.0